Extract source-agnostic TaskWrapper ABC from OpenMLTaskWrapper
benchmark/task/wrapper.py now owns the runtime task interface the run
engine consumes: split-index arithmetic, train/test split assembly
(incl. lazy loading + subsampling), metric-error computation, the light
dtype flags, and a default get_validation_metadata. Subclasses provide
the source-specific hooks: _load_data, task_id, get_split_dimensions,
get_split_indices.

OpenMLTaskWrapper becomes the OpenML-backed subclass (data/splits
delegated to the OpenML task object). The experiment engine
(ExperimentRunner, Experiment, _LazyTask) is retyped against the ABC,
and task_cache_scope now uses the public TaskWrapper.has_text.

No behavior change.
