+++ packages/tabarena/src/tabarena/benchmark/task/__init__.py
@@ -4,5 +4,6 @@ from tabarena.benchmark.task.user_task import (
     UserTask,
     from_sklearn_splits_to_user_task_splits,
 )
+from tabarena.benchmark.task.wrapper import TaskWrapper
 
-__all__ = ["UserTask", "from_sklearn_splits_to_user_task_splits"]
+__all__ = ["TaskWrapper", "UserTask", "from_sklearn_splits_to_user_task_splits"]
